% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{A title}
\usage{
add(a, b)
}
\arguments{
\item{a, b}{Arguments.}
}
\description{
A title
}
\section{options}{

\itemize{
\item \code{item1}:
\mjtdeqn{
F(t) = \left\lbrace\begin{array}{ll}
\alpha(t) f_{L}(t) + (1 - \alpha(t)) f_{C}(t), & t_{L} < t < t_{C}, \cr
\beta(t) f_{C}(t) + (1 - \beta(t)) f_{R}(t), & t_{C} < t < t_{R}. \end{array}\right.}{\%
F(t) = \begin{cases}
\alpha(t) f_{L}(t) + (1 - \alpha(t)) f_{C}(t), & t_{L} < t < t_{C}, \cr
\beta(t) f_{C}(t) + (1 - \beta(t)) f_{R}(t), & t_{C} < t < t_{R}. \end{cases}}{}
\item item2
}

\loadmathjax{}
}

