% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello-package.R
\docType{package}
\name{Hello-package}
\alias{Hello}
\alias{Hello-package}
\title{Hello: Say hello in different languages}
\description{
This is a toy package that contains a function that says hello when called.
}
\author{
\strong{Maintainer}: Anouar El ghouch \email{anouar.elghouch@uclouvain.be}

}
\keyword{internal}
