% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(who, lang = "EN", LangData = Hello::language)
}
\arguments{
\item{who}{a \code{character} vector of length 1 that specifies the name of the
person to whom the message is adressed.}

\item{lang}{a \code{character} vector of length 1 that specifies the preferred
language. Default to "EN" for English. Other possible values include "FR"
for French, "IT" for Italian, "ES" for Spanish, or "DE" for German. Case is
ignored."}

\item{LangData}{optional data.frame with two columns each of mode character. The
first column gives the language codes and the second column gives the
corresponding "hello" word. Default to language. see ?language}
}
\value{
a \code{character} vector with a personalized \emph{"hello"} message.
}
\description{
\code{hello} says \emph{"hello"} in the user-specified language. The user is asked to
give her/his name so that the hello massage gets personalized.
}
\examples{
hello("james")
hello("Amelia", "Es")

}
