% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\name{hello2}
\alias{hello2}
\title{hello2}
\usage{
hello2()
}
\arguments{
\item{a, b}{Arguments.}
}
\description{
\loadmathjax{}
A description.
}
\details{
An irrelevant equation:
\mjtdeqn{F(t) = \left\lbrace\begin{array}{ll} \alpha(t) f_{L}(t) + \lbrack 1 - \alpha(t) \rbrack f_{C}(t)\,, & t_{L} < t < t_{C}\,, \cr \beta(t) f_{C}(t) + \lbrack 1 - \beta(t) \rbrack f_{R}(t)\,, & t_{C} < t < t_{R}\,. \end{array}\right.}{%
          F(t) = \begin{cases} \alpha(t) f_{L}(t) + \lbrack 1 - \alpha(t) \rbrack f_{C}(t)\,, & t_{L} < t < t_{C}\,, \cr \beta(t) f_{C}(t) + \lbrack 1 - \beta(t) \rbrack f_{R}(t)\,, & t_{C} < t < t_{R}\,. \end{cases}}{%
          ... a plain text translation ...}
}
