% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_detect_language.R
\name{hello_detect_language}
\alias{hello_detect_language}
\title{Detect language}
\usage{
hello_detect_language(greeting)
}
\arguments{
\item{greeting}{Input is a character vector}
}
\value{
character vector
}
\description{
'hello_detect_language()' detects the language of basic greetings in English, French, Spanish, German and Italian
}
\examples{
hello_detect_language('bonjour')
hello_detect_language('hi')
}
