% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_message}
\alias{hello_message}
\title{Generate a greeting}
\usage{
hello_message(x = "world")
}
\arguments{
\item{x}{The name to greet.}
}
\value{
A greeting message.
}
\description{
Generate a greeting
}
