% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/io_dist_plot.R
\name{io_dist_plot}
\alias{io_dist_plot}
\title{io_dist_plot}
\usage{
io_dist_plot(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\value{
a renderPlot result
}
\description{
Shiny server side function to construct a distribution plot
}

