% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4-package.R
\docType{package}
\name{lab4-package}
\alias{lab4}
\alias{lab4-package}
\title{lab4: Hello: Say hello in different languages!}
\description{
This is a toy packages containing a function that says hello when called.
}
\author{
\strong{Maintainer}: Emilie Do \email{thuy.l.do@student.uclouvain.be} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
