% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{The \code{language} dataset}
\format{
\subsection{}{

This is a data frame with 5 rows and 2 columns:
\itemize{
\item \code{code}: language code: "en", "fr", "it", "es", "de", for English, French, Italian, Spanish, and German.
\item \code{hello}: word "hello" in the language corresponding to the code
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of}

\url{https://translate.google.com}
}
\usage{
language
}
\description{
\code{language} provides the word "hello" in different languages.
}
\section{Author(s)}{
Anouar El Ghouch \href{mailto:anouar.elghouch@uclouvain.be}{anouar.elghouch@uclouvain.be}
}

\keyword{datasets}
