% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuns.R
\name{mfuns}
\alias{mfuns}
\alias{fun1}
\alias{fun2}
\alias{fun3}
\title{Multiple functions}
\usage{
fun1(x)

fun2(x, y)

fun3(x, y, z)

fun2(x, y)

fun3(x, y, z)
}
\arguments{
\item{x}{a common argument}

\item{y}{fun2 specific}

\item{z}{fun3 specific}
}
\description{
multi-function in an Rd file.
}
