% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{monde}
\alias{monde}
\title{Say hello to the world with a cow}
\usage{
monde(...)
}
\arguments{
\item{...}{Arguments passed to \code{cowsay::say}}
}
\description{
Say hello, in some language, with a cow.  This version uses an
implicit language argument
}
