% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1labpkg-package.R
\docType{package}
\name{my1labpkg-package}
\alias{my1labpkg}
\alias{my1labpkg-package}
\title{my1labpkg: First Lab Package}
\description{
I don't know yet what this package does. It's just an exercise. 
 I can write many things, but I have to leave 2 spaces at the beginning of a new line.
}
\author{
\strong{Maintainer}: Eleonore Beckers \email{eleonore.beckers@uclouvain.be}

}
\keyword{internal}
