% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_description.R
\name{my_description}
\alias{my_description}
\title{A function to quickly describe yourself}
\usage{
my_description(nickname, age, animal)
}
\arguments{
\item{nickname}{the name you like to be called}

\item{age}{how old you are}

\item{animal}{your favorite animal}
}
\description{
A function to quickly describe yourself
}
\examples{
my_description("Jo", 26, "moose")
}
