% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mat.R
\docType{data}
\name{pois_mat}
\alias{pois_mat}
\title{Sample data document}
\format{An object of matrix with 100 rows and 100 colmns.}
\usage{
data(pois_mat)
}
\description{
Data from a 'rpois()'.
}
\examples{
data(pois_mat)
brk <- pretty(pois_mat, n = nclass.Sturges(pois_mat))
hist(pois_mat, breaks = brk)
}
\references{
\href{https://github.com/shkonishi/hello/tree/master/man}{man}
}
\keyword{datasets}
