% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton-package.R
\docType{package}
\name{skeleton-package}
\alias{skeleton}
\alias{skeleton-package}
\title{skeleton: Hello}
\description{
This is a toy package containing a function that says hello when called.
}
\author{
\strong{Maintainer}: Colleen Peeters \email{colleen.peeters@student.uclouvain.be} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
