% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTO-package.R
\docType{package}
\name{TESTO-package}
\alias{TESTO}
\alias{TESTO-package}
\title{TESTO: A dumb package}
\description{
Toy package.
}
\author{
\strong{Maintainer}: Robert Paulus \email{first.last@example.com}

}
\keyword{internal}
