% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_bins.R
\name{ui_bins}
\alias{ui_bins}
\title{ui_bins}
\usage{
ui_bins()
}
\value{
an input element
}
\description{
A shiny user interface fragment to select n of bins.
}
\seealso{
Other shiny_ui: \code{\link{ui_dist_plot}}
}

