% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ui_dist_plot.R
\name{ui_dist_plot}
\alias{ui_dist_plot}
\title{ui_dist_plot}
\usage{
ui_dist_plot()
}
\value{
an output element
}
\description{
A shiny user interface fragment to show a distribution plot.
}
\seealso{
Other shiny_ui: \code{\link{ui_bins}}
}

