% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{world}
\alias{world}
\title{Say hello to the world with a cow}
\usage{
world(language = "en", ...)
}
\arguments{
\item{language}{Language to translate into}

\item{...}{Arguments passed to \code{cowsay::say}}
}
\description{
Say hello, in some language, with a cow.  This version uses an
explicit language argument
}
