% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello-package.R
\docType{package}
\name{Hello-package}
\alias{Hello}
\alias{Hello-package}
\title{Hello: Hello: Say hello in different languages!}
\description{
This a toy package containing a function that says hello when called.
}
\author{
\strong{Maintainer}: Christophe Ullrich \email{christophe.ullrich@student.uclouvain.be}

}
\keyword{internal}
