% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello-package.R
\docType{package}
\name{Hello-package}
\alias{Hello}
\alias{Hello-package}
\title{Hello: Say hello in different languages!}
\description{
This is a toy package containing a function that says hello when called.
}
\author{
\strong{Maintainer}: Mathieu Lecart \email{mathieu.lecart@uclouvain.be}

}
\keyword{internal}
