% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello-package.R
\docType{package}
\name{Hello-package}
\alias{Hello}
\alias{Hello-package}
\title{Hello: Say hello in different languages!}
\description{
This is a toy package...
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
