\name{hello-package}
\alias{hello-package}
\alias{hello}
\docType{package}
\title{
Have a say(x) function to print x to screen.
}
\description{
This is an example to create R package.
}
\details{
\tabular{ll}{
Package: \tab hello\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-03\cr
License: \tab What license is it under?\cr
}
Just use simply
}
\author{
Me.

Maintainer: Joowhi Lee (joohwi@cs.unc.edu)
}
\references{
None
}
None
\keyword{ package }
\seealso{
None
}
\examples{
}
