% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(who, lang = "en", langdata = language)
}
\arguments{
\item{who}{Character string name of recipient}

\item{lang}{language code from langdata default "en", "fr" "it" "es" "de"}

\item{langdata}{dataframe first column named code for language code string second column named hello for hello string in corresponding language}
}
\value{
\code{'Hello, who!'} but in the specified language
}
\description{
\code{hello} says \emph{'Hello, world!'}
}
