% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(who, lang = "en", langData = Hello::language)
}
\arguments{
\item{who}{specifies the name of the person to whom the “hello”
message is addressed. This must be a character vector of length 1.}

\item{lang}{character vector, length 1,
specifies the user preferred language. This can be “EN” (the default value)
for English, “FR” for French, “IT” for Italian, “ES” for Spanish,
or “DE” for German. Upper and lower cases should be interpreted as identical.}

\item{langData}{an optional data.frame with two columns.
The first column gives the language codes and the second column gives
the corresponding “hello” word.
The default value of LangData is set to the dataset language.
see \code{?language}}
}
\value{
a \code{character} vector with a personalized "hello" message
}
\description{
\code{hello} says \emph{"Hello, world!"} in chosen language
}
\examples{
hello("james")
hello("amelia","Es")

}
