% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{hello}
\alias{hello}
\alias{hello-package}
\title{hello: A minimal test R-Package}
\description{
'hello' is an minimal package example containing a simple "Hello World"
script. This was designed to test out issues had with R-package developer
tools, such as Rstudio's build commands.
}

