% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(who="world", lang="en",langdata)
}
\arguments{
\item{who}{specifies the name of the person to whom the “hello” message is addressed. This must be a character vector of length 1.}

\item{lang}{a character vector of length 1 that specifies the user preferred language. This can be "EN" (the default value) for English, "FR" for French, “IT” for Italian, “ES” for Spanish, or “DE” for German.
Case should be ignored the code should consider ‘a’ and ‘A’ to be the same characters. If an undefined language is chosen, the function must show a message about that (see below).}

\item{langdata}{an optional data.frame with two columns. The first column gives the language codes and the second column gives the corresponding “hello” word. The default value should be set to the data.frame language defined above.}
}
\value{
a character vector with a personalized "hello" message
}
\description{
\code{hello} says \emph{'Hello, world!'}
}
\examples{
hello("James")
hello("Amelia")

}
