% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(name = "you")
}
\arguments{
\item{name}{A character vector of one element which is the name to concatenate to Hello}
}
\value{
A character vector of one element which contains Hello followed by content of the parameter
}
\description{
A function that returns Hello followed by the named passed as a parameter
}
\examples{
hello("Hubert")

}
