% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(who, lang = "EN", LangData = Hello::language)
}
\arguments{
\item{who}{a `character` vector of length 1 that specifies the name of the person to whom the message is addressed.}

\item{lang}{a `character` vector of length 1 that specifies the preferred language. Default to "EN" for English. Other possible values include "FR" for French, "IT"  for Italian, "ES" for Spanish, or "DE" for German. Case is ignored.}

\item{LangData}{an optional data.frame with two columns each of mode `character`. The first column gives the language codes and the second column gives the corresponding "hello" word. Default to `language`.

see `?language`}
}
\value{
a `character` vector with a personalized _"hello"_ message.
}
\description{
`hello` says _"hello"_ in the user-specified language. The user is asked to give her/his name so that the hello message gets personalized.
}
\examples{
hello("James")
hello("Amelia", "Es")

}
