% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{a string; default is 'world'}
}
\value{
character a string; default is 'Hello, world!'
}
\description{
The typical hello world example.
}
\examples{
# Example usage of hello()

hello()

hello("all") 
}
\author{
Reinhard Simon
}
\seealso{
Other example: \code{\link{hello_world}}
}

