% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello to someone}
\usage{
hello(who)
}
\arguments{
\item{who}{character(1) The name of the person to say hello to.}
}
\value{
character(1) The greeting.
}
\description{
This command can be invoked to say hello to a person.
}
\examples{
hello("Han")

}
