% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
hello(n)
}
\arguments{
\item{n}{Number of replication}
}
\value{
Character vector, length of n
}
\description{
Prints 'Hello, world!'.
}
\examples{
hello(3)
}
