% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(who,lang="EN",LangData=language)
}
\arguments{
\item{who}{a \code{character} vector of length 1 that specifies the name of the person}

\item{lang}{a \code{character} vector of length 1}

\item{LangData}{an optional data.frame with two colmns each of mode \code{character}.}
}
\value{
a \code{character} vector with a personalized \emph{"hello"} message.
}
\description{
\code{hello} says \emph{"Hello"} in the user-specified language. The user is asked to give her/his name so that the hello message gets personalized.
}
\examples{
hello("James")
hello("Amélie","Es")

}
