% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{hello}
\alias{hello}
\title{Title
Hello World}
\usage{
hello(who, lang = "EN", langData = Hello::language)
}
\arguments{
\item{who}{a \code{character} vector of length 1, ..}

\item{lang}{a \code{character} vector of}

\item{langData}{a vector see \code{?language}}
}
\value{
a \code{character} vector with person
}
\description{
\code{hello} says \emph{"Hello"} in the user specified language. the user is asked to
give her/his name so that the hello message gets personalized.
}
\examples{
# Say hello
hello("James")
hello("Amelia", "Es")

}
