% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4-package.R
\docType{package}
\name{lab4-package}
\alias{lab4}
\alias{lab4-package}
\title{lab4: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alexis Franco \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
