% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{Language}
\format{
A data frame with 5 rows and 2 columns:
\describe{
\item{code}{language code en fr it es de for language}
\item{hello}{word "Hello" in the language corresponding to the code}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
language
}
\description{
The data frame \code{language} provides the word "hello" in different languages
}
\keyword{datasets}
