% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{Language}
\format{
A \code{data.frame} with 5 rows and 2 columns:
\describe{
\item{code}{language code: "en","fr","it","es","de", for English, French, Italian, Spanish, and German.}
\item{hello}{word "hello" in thhe language corresponding to the code.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}
\url{https://translate.google.com/}
}
\usage{
language
}
\description{
The data.frame \code{language} provides the word "hello" in different languages.
}
\author{
Colleen Peeters
}
\keyword{datasets}
