% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{Language}
\format{
A data.frame with 5 rows and columns:
\describe{
\item{code:}{language code: "en", "fr", etc}
\item{hello:}{word hello in the language}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
language
}
\description{
The data.frame language provides the world "hello"
in the different languages
}
\keyword{datasets}
