% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{The `language` dataset}
\format{
This is a `data.frame` with `r nrow(language)` rows and `r ncol(language)` columns:

* `code`: language code: "en","fr","it","es","de", for
          English, French, Italian, Spanish, and German.

* `hello`: word "hello" in the language corresponding to the code.
}
\source{
<https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes>

<https://translate.google.com>
}
\usage{
language
}
\description{
`language` data provides the word "hello" in different languages.
}
\author{
Anouar El Ghouch  <anouar.elghouch@uclouvain.be>
}
\keyword{datasets}
