% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\docType{data}
\name{language}
\alias{language}
\title{The \code{language} dataset}
\format{
This is a \code{data.frame} with 5 rows and 2 columns:
\itemize{
\item \verb{code:} language code: "en","fr","it","es","de", for English, French, Italian, Spanish, and German.
\item \verb{hello:} word "hello" in the language corresponding to the code.
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}

\url{https://translate.google.com/}
}
\usage{
language
}
\description{
\code{language} data provides the word "hello" in different languages.
}
\author{
Lucas Elvira \url{lucas.elvira@student.uclouvain.be}
}
\keyword{datasets}
