% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{The \code{language} dataset}
\format{
This is a \code{data.frame} with 5 rows and 2 columns:

*\code{code}: language code: "en","fr","it","es","de", for English, French, Italian, Spanish, and German
*\code{hello}: word "hello" in the language corresponding to the code.
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_ISO_639-&_codes}

\url{https://translate.google.com}
}
\usage{
language
}
\description{
\code{languages} data provides the world "hello" in different languages
}
\author{
Anouar El Ghouch \href{mailto:anouar.elghouch@uclouvain.be}{anouar.elghouch@uclouvain.be}
}
\keyword{datasets}
