% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{Language}
\format{
A \code{data.frame} with 5 rows and 2 columns:
\describe{
\item{\code{code}}{language code: "en", "fr", "it", "es", "de", for English, French, Italian, Spanish and German.}
\item{\code{hello}}{word "hello" in the language correspondig to the code}
}
}
\source{
{\url{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}

\url{https://translate.google.com/}}
}
\usage{
language
}
\description{
The dataframe \code{language} provides the word "hello" in different languages.
}
\keyword{datasets}
