% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{The ´language´ dataset}
\format{
This is a ´data.frame´ with 5 rows and 2 columns
\itemize{
\item code : language stuff
\item hello : the word hello in different language
}
}
\source{
\url{https://google.com}
}
\usage{
language
}
\description{
´language´ data is very cool
}
\author{
Robert Paulus \href{mailto:Robert.Paulus@student.uclouvain.be}{Robert.Paulus@student.uclouvain.be}
}
\keyword{datasets}
