% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{The \code{language} dataset}
\format{
This is a \code{data.frame} with 5 rows and 2 columns:
\itemize{
\item \code{code}
\item \code{hello}
}
}
\source{
\url{https://translate.google.com}
}
\usage{
language
}
\description{
\code{language} data provides the word "hello" in different languages
}
\author{
Stephane De Bruille \url{sdebruille@hotmail.com}
}
\keyword{datasets}
