% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{The \code{language} dataset}
\format{
This is a \code{data.frame} with 5 rows and 2 columns :
\itemize{
\item \code{code} : language code : "en","fr","it","es","de", for English, French, Italian, Spanish and German
\item \code{hello} : word "Hello" in the language corresponding to the code
}
}
\source{
\url{https://en.wikipedia.org}
}
\usage{
language
}
\description{
\code{language} data provides the word "hello" in different languages.
}
\author{
Emilie Do
}
\keyword{datasets}
