% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{The \code{language} dataset}
\format{
This is a \code{data.frame} with 5 rows and 2 columns:
\itemize{
\item{\verb{code:}}{language code: 'en', 'fr','it', 'es', 'de', for English, French, Italian, Spanish, and German.}
\item{\verb{hello:}}{word "hello in the language corresponding to the code}
}
}
\source{
\url{https://www.google.com/search?client=firefox-b-d&q=google+translate}

\url{https://www.google.com/search?client=firefox-b-d&q=google+translate}
}
\usage{
language
}
\description{
\code{language} data provides the word "hello" in different languages.
}
\author{
Tom Van Zeebroeck \url{tom.vanzeebroeck@uclouvain.be}
}
\keyword{datasets}
